requirements.txt
pygame>=2.0.0
google-genai>=0.1.0
